Traktor: Various fixes; mostly Linux/GCC related. New BinaryInclude binary which checks file timestamps prior to converting files.
[git-p4: depot-paths = "//depot/traktor/": change = 3029]

+++ code/Flash/Acc/AccTextureCache.h
@@ -41,7 +41,7 @@ public:
 
 private:
 	Ref< render::IRenderSystem > m_renderSystem;
-	std::map< uint32_t, Ref< render::ITexture > > m_cache;
+	std::map< uint64_t, Ref< render::ITexture > > m_cache;
 };
 
 	}
